# swagger.yml
openapi: 3.0.0
info:
  title: "Trail Management API"
  description: "API for user authentication, trail management, and feature management"
  version: "1.0.0"
servers:
  - url: "/api"
paths:
...
  /trails:
    get:
      tags:
        - Trails
      summary: "Retrieve all trails"
      description: "Fetch a list of all trails from the database."
      operationId: trails.read_all
      responses:
        "200":
          description: "List of trails retrieved successfully"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrailWithFeatures"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"



    post:
      tags:
        - Trails
      summary: "Create a new trail"
      description: >
        Create a trail linked to the logged-in user. The payload supports adding waypoints and features at the time of creation.
      operationId: trails.create_trail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrailWithFeatures"
            example:
              trail_name: "Ocean View Trail"
              trail_summary: "A beautiful trail with ocean views."
              trail_description: "This trail is known for its scenic ocean views and gentle terrain."
              difficulty: "Easy"
              location: "Cornwall, UK"
              length: 5.5
              elevation_gain: 150
              route_type: "Loop"
              waypoints:
                pt1:
                  lat: 50.1234
                  long: -5.6789
                  desc: "Start of the trail"
                pt2:
                  lat: 50.124
                  long: -5.6795
                  desc: "Viewpoint overlooking the ocean"
                pt3:
                  lat: 50.125
                  long: -5.68
                  desc: "End of the loop"
              features:
                - feature_name: "Waterfall"
                - feature_name: "ETC"
      responses:
        "201":
          description: "Trail created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailWithFeatures"
        "400":
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trails/{trail_id}:
    get:
      tags:
        - Trails
      summary: Retrieve a trail by ID
      description: >
        Fetch a specific trail by its ID. This endpoint requires the user to be authenticated 
        with a valid session. Role-based access control ensures only authorized roles can access this endpoint.
      operationId: trails.read_by_id
      parameters:
        - name: trail_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
            description: The ID of the trail to retrieve.
      responses:
        "200":
          description: Trail retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  trail_name:
                    type: string
                    example: "Ocean View Trail"
                  difficulty:
                    type: string
                    example: "Easy"
                  location:
                    type: string
                    example: "Cornwall, UK"
                  length:
                    type: number
                    example: 5.5
                  elevation_gain:
                    type: number
                    example: 150
        "403":
          description: Forbidden. User lacks permission to access this endpoint.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Forbidden. You do not have permission to access this resource."
        "401":
          description: Unauthorized. User is not logged in.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "User is not logged in."
        "404":
          description: Trail not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Trail not found."

                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Trails
      summary: Update a trail by ID
      description: >
        Update an existing trail's details, including waypoint updates and feature management. 
        Use `features.add` to add features to the trail and `features.remove` to remove them.
      operationId: trails.update_trail
      parameters:
        - name: trail_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Fields to update for the trail
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrail"
            example:
              trail_name: "Updated Name"
              trail_summary: "An updated trail summary."
              trail_description: "An updated trail description."
              difficulty: "Updated"
              location: "Updated, UK"
              length: 50.0
              elevation_gain: 200.0
              route_type: "Updated"
              waypoints:
                pt1:
                  lat: 50.1231
                  long: -5.6718
                  desc: "Updated Desc"
                pt2:
                  lat: 50.124
                  long: -5.679
                  desc: "Updated Desc"
                pt3:
                  lat: 50.125
                  long: -5.680
                  desc: "Updated Desc"
              features:
                add: ["Scenic update", "Historic Landmark"]
                remove: ["Mountain Pass", "Large Waterfall"]

      responses:
        "200":
          description: Trail updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrailWithFeatures"
        "400":
          description: Validation error or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Trail not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Trails
      summary: "Delete a trail"
      description: "Delete a trail by its ID and remove its links to features."
      operationId: trails.delete_trail
      parameters:
        - name: trail_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: "Trail deleted successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Trail deleted successfully."
        "404":
          description: "Trail not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trails/{trail_id}/features:
    post:
      tags:
        - Trails
      summary: "Add a feature to a trail"
      description: "Add a feature to a trail. If the feature does not exist, it will be created."
      operationId: trails.add_feature_to_trail
      parameters:
        - name: trail_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feature_name:
                  oneOf:
                    - type: string
                      example: "Scenic Waterfall"
                    - type: array
                      items:
                        type: string
                      example: ["Scenic Waterfall", "Viewpoint"]
      responses:
        "201":
          description: "Feature added successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature added successfully."
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Trails
      summary: "Remove a feature from a trail"
      description: "Remove a feature from a trail by deleting the association in the TrailFeature table."
      operationId: trails.remove_feature_from_trail
      parameters:
        - name: trail_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feature_name:
                  oneOf:
                    - type: string
                      example: "Scenic Overlook"
                    - type: array
                      items:
                        type: string
                      example: ["Scenic Overlook", "Waterfall"]
      responses:
        "200":
          description: "Feature successfully removed from the trail."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature 'Waterfall' successfully removed from trail ID 1."
        "400":
          description: "Validation error or feature not linked to trail."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Trail or feature not found."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal server error."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  #################### Feature Endpoints ####################
  /features:
    get:
      summary: Get All Features
      description: Fetch all features from the database and return them as a JSON response.
      operationId: features.read_all_features
      tags:
        - Features
      responses:
        '200':
          description: A list of features.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feature'
        '500':
          description: An error occurred on the server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Add a New Feature
      description: Add a new feature to the database.
      operationId: features.add_feature
      tags:
        - Features
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feature_name:
                  type: string
                  example: "Add new Feature"
      responses:
        '201':
          description: Feature successfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Feature 'Waterfall' successfully added."
                  feature:
                    type: object
                    properties:
                      feature_name:
                        type: string
                        example: "Waterfall"
        '400':
          description: Validation error or feature already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    TrailWithFeatures:
      type: object
      properties:
        trail_name:
          type: string
        trail_summary:
          type: string
        trail_description:
          type: string
        difficulty:
          type: string
        location:
          type: string
        length:
          type: number
          format: float
        elevation_gain:
          type: number
          format: float
        route_type:
          type: string
        waypoints:
          type: object
          properties:
            pt1:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
            pt2:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
            pt3:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                long:
                  type: number
                  format: float
                desc:
                  type: string
        features:
          type: array
          items:
            type: object
            properties:
              feature_name:
                type: string
    UpdateTrail:
      type: object
      properties:
        trail_name:
          type: string
          example: "Updated Trail Name"
        difficulty:
          type: string
          example: "Moderate"
        waypoints:
          type: object
          properties:
            pt1:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                  example: 50.123
                long:
                  type: number
                  format: float
                  example: -5.678
                desc:
                  type: string
                  example: "Updated Start Point"
            pt2:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                  example: 50.124
                long:
                  type: number
                  format: float
                  example: -5.679
                desc:
                  type: string
                  example: "Updated Mid Point"
            pt3:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                  example: 50.125
                long:
                  type: number
                  format: float
                  example: -5.680
                desc:
                  type: string
                  example: "Updated End Point"
        features:
          type: object
          properties:
            add:
              type: array
              items:
                type: string
              example: ["Scenic Waterfall", "Viewpoint"]
            remove:
              type: array
              items:
                type: string
              example: ["Old Path", "Rocky Path"]
    Feature:
      type: object
      properties:
        feature_name:
          type: string
          example: "Waterfall"            
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred."